fix: 별점 관련 훅 분리 및 fillColor 추가
