+++ src/pages/main/hooks/starRate.tsx
@@ -1,0 +1,30 @@
+import Icon, { IconValues } from "@/components/icon";
+import { Colors } from "@/styles/colorPalette";
+
+
+export const starRate = (reviewAverage: number) => {
+    const icons = [];
+    for (let i = 0; i < 5; i++) {
+      let fillColor: Colors = 'black';
+      let value: IconValues = 'review';
+
+      if (i < Math.floor(reviewAverage)) {
+        fillColor = 'white';
+      } else if (i === Math.floor(reviewAverage) && reviewAverage % 1 !== 0) {
+        value = 'halfreview'
+        fillColor = 'white';
+      }
+
+      icons.push(
+        <Icon
+          key={i}
+          value={value}
+          strokeColor="white"
+          fillColor={fillColor}
+          size={30}
+          $active={false}
+        />,
+      );
+    }
+    return icons;
+};
